requirements.txt
asgiref==3.4.1
certifi==2020.6.20
Django==3.2.9
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.0
wincertstore==0.2
mysqlclient
djangorestframework
django-filter
markdown
djangorestframework_simplejwt